LayoutTests/accessibility/mac/AOM-number-properties.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../resources/accessibility-helper.js"></script>
</head>
<body>

<div role="slider" id="slider" aria-valuenow="5"></div>

<div id="grid" role="grid">
  <div role="rowgroup">
    <div role="row">
      <span id="cell" role="columnheader">First Name</span>
    </div>
  </div>
</div>

<h role="heading" id="heading">
<div role="list">
<div role="listitem" id="listitem"></div>
</div>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests number type Accessibility Object Model properties.");
    if (window.accessibilityController) {
        var node;
        var axNode;

        testClearNumberProperty();
        testRangeProperties();
        testTableProperties();
        testLevel();
        testList();
    }

    function testClearNumberProperty() {
        debug("\nTest AOM number properties can be cleared.");
        node = document.getElementById("slider");
        axNode = accessibilityController.accessibleElementById("slider");
        // Number value should be null if not set.
        shouldBeEqualToString("axNode.stringValue", "AXValue: 5");
        shouldBeNull("node.accessibleNode.valueNow");
        // Set the AOM property
        node.accessibleNode.valueNow = 6;
        shouldBeEqualToString("axNode.stringValue", "AXValue: 6");
        // Clear the AOM property
        node.accessibleNode.valueNow = null;
        shouldBeNull("node.accessibleNode.valueNow");
        shouldBeEqualToString("axNode.stringValue", "AXValue: 5");
    }

    function testRangeProperties() {
        debug("\nValueMax, ValueMin and ValueNow support");
        node = document.getElementById("slider");
        axNode = accessibilityController.accessibleElementById("slider");
        shouldBe("axNode.maxValue", "100");
        shouldBe("axNode.intValue", "5");
        shouldBe("axNode.minValue", "0");
        node.accessibleNode.valueMax = 20;
        node.accessibleNode.valueNow = 15;
        node.accessibleNode.valueMin = 10;
        shouldBe("axNode.maxValue", "20");
        shouldBe("axNode.intValue", "15");
        shouldBe("axNode.minValue", "10");
    }

    function testTableProperties() {
        debug("\nTable related properties support");
        // colCount and rowCount
        node = document.getElementById("grid");
        axNode = accessibilityController.accessibleElementById("grid");
        shouldBe("axNode.numberAttributeValue('AXARIAColumnCount')", "0");
        shouldBe("axNode.numberAttributeValue('AXARIARowCount')", "0");
        node.accessibleNode.colCount = 16;
        node.accessibleNode.rowCount = 30;
        shouldBe("axNode.numberAttributeValue('AXARIAColumnCount')", "16");
        shouldBe("axNode.numberAttributeValue('AXARIARowCount')", "30");
        
        // colSpan and rowSpan
        node = document.getElementById("cell");
        axNode = accessibilityController.accessibleElementById("cell");
        shouldBe("axNode.columnIndexRange()", "'{0, 1}'");
        shouldBe("axNode.rowIndexRange()", "'{0, 1}'");
        node.accessibleNode.colSpan = 2;
        node.accessibleNode.rowSpan = 3;
        shouldBe("axNode.columnIndexRange()", "'{0, 2}'");
        shouldBe("axNode.rowIndexRange()", "'{0, 3}'");
        
        // colIndex and rowIndex
        shouldBe("axNode.numberAttributeValue('AXARIAColumnIndex')", "-1");
        shouldBe("axNode.numberAttributeValue('AXARIARowIndex')", "-1");
        node.accessibleNode.colIndex = 1;
        node.accessibleNode.rowIndex = 2;
        shouldBe("axNode.numberAttributeValue('AXARIAColumnIndex')", "1");
        shouldBe("axNode.numberAttributeValue('AXARIARowIndex')", "2");
    }

    function testLevel() {
        debug("\nLevel support");
        node = document.getElementById("heading");
        axNode = accessibilityController.accessibleElementById("heading");
        // The implicit value of aria-level is 2 for the heading role.
        shouldBe("axNode.intValue", "2");
        node.accessibleNode.level = 6;
        shouldBe("axNode.intValue", "6")
    }

    function testList() {
        debug("\nPosInSet and SetSize support");
        node = document.getElementById("listitem");
        axNode = accessibilityController.accessibleElementById("listitem");
        node.accessibleNode.posInSet = 9;
        node.accessibleNode.setSize = 10;
        shouldBe("axNode.numberAttributeValue('AXARIAPosInSet')", "9");
        shouldBe("axNode.numberAttributeValue('AXARIASetSize')", "10");
    }
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
