LayoutTests/accessibility/mac/AOM-number-properties-expected.txt

First Name
This tests number type Accessibility Object Model properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Test AOM number properties can be cleared.
PASS axNode.stringValue is "AXValue: 5"
PASS node.accessibleNode.valueNow is null
PASS axNode.stringValue is "AXValue: 6"
PASS node.accessibleNode.valueNow is null
PASS axNode.stringValue is "AXValue: 5"

ValueMax, ValueMin and ValueNow support
PASS axNode.maxValue is 100
PASS axNode.intValue is 5
PASS axNode.minValue is 0
PASS axNode.maxValue is 20
PASS axNode.intValue is 15
PASS axNode.minValue is 10

Table related properties support
PASS axNode.numberAttributeValue('AXARIAColumnCount') is 0
PASS axNode.numberAttributeValue('AXARIARowCount') is 0
PASS axNode.numberAttributeValue('AXARIAColumnCount') is 16
PASS axNode.numberAttributeValue('AXARIARowCount') is 30
PASS axNode.columnIndexRange() is '{0, 1}'
PASS axNode.rowIndexRange() is '{0, 1}'
PASS axNode.columnIndexRange() is '{0, 2}'
PASS axNode.rowIndexRange() is '{0, 3}'
PASS axNode.numberAttributeValue('AXARIAColumnIndex') is -1
PASS axNode.numberAttributeValue('AXARIARowIndex') is -1
PASS axNode.numberAttributeValue('AXARIAColumnIndex') is 1
PASS axNode.numberAttributeValue('AXARIARowIndex') is 2

Level support
PASS axNode.intValue is 2
PASS axNode.intValue is 6

PosInSet and SetSize support
PASS axNode.numberAttributeValue('AXARIAPosInSet') is 9
PASS axNode.numberAttributeValue('AXARIASetSize') is 10
